FastAPI/requirements.txt
SQLAlchemy==2.0.17
PyMySQL==1.0.2
mysql-connector-python==8.0.29
mariadb==1.0.11
pydantic==1.9.0
PyJWT==2.8.0
Werkzeug==2.3.7
cryptography==41.0.4
fastapi==0.99.1
crypto==1.4.1
pycryptodome==3.19.0
uvicorn==0.24.0